refactor: move verification logic from VerificationController to AuthController and update route
