+++ turnomaster/app/Http/Controllers/AuthController.php
@@ -86,4 +86,28 @@ class AuthController extends Controller
         return redirect('/');
     }
     
+    public function verify($code)
+    {
+        $verificationCode = VerificationCode::where('code', $code)->first();
+
+        if (!$verificationCode) {
+            return redirect()->route('email.handler')->with('error', 'El código de verificación es inválido o ya ha sido utilizado.');
+        }
+
+        if ($verificationCode->isExpired()) {
+            return redirect()->route('email.handler')->with('error', 'El código de verificación ha expirado.');
+        }
+
+        $user = $verificationCode->user;
+        $user->activated_account = true;
+        $user->email_verified_at = Carbon::now();
+        $user->save();
+
+        $verificationCode->delete();
+
+        $userName = ucwords(strtolower($user->name));
+
+        return redirect('/')->with('success', 'Su cuenta ha sido verificada exitosamente.')->with('verified', true)->with('userName', $userName);
+    }
+
 }
